fixed: Some missing cleanups from r23060, and a possible crash when using fullscreen sliders.
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/linuxport/XBMC@23069 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ guilib/GUIVisualisationControl.cpp
@@ -401,7 +401,7 @@ bool CGUIVisualisationControl::OnMessage(CGUIMessage &message)
   else if (message.GetMessage() == GUI_MSG_VISUALISATION_ACTION)
   {
     CAction action;
-    action.id = (int)message.GetParam1();
+    action.id = message.GetParam1();
     return OnAction(action);
   }
   else if (message.GetMessage() == GUI_MSG_PLAYBACK_STARTED)
